Сделал cache invalidation для showcase-groups при изменении товара, который в данный момент находится в одной из групп

+++ AdminPanel/MediatorHandlers/Products/DeleteProductCommand.cs
@@ -21,8 +21,12 @@ public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand>
 
     public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
-        var product = await _context.Products.FirstOrDefaultAsync(x => x.Id == request.ProductId, cancellationToken);
+        var product = await _context.Products
+            .Include(x => x.ShowcaseGroups)
+            .FirstOrDefaultAsync(x => x.Id == request.ProductId, cancellationToken);
         if (product is null) throw new HttpRequestException($"No product with was found with id {request.ProductId}");
+
+        if(product.ShowcaseGroups.Count > 0) throw new HttpRequestException($"Product is in showcase group. Can not delete it");
 
         var priceHistories = await _context.PriceHistories.Where(x => x.ProductId == request.ProductId).ToListAsync(cancellationToken);
         if (priceHistories.Any()) _context.PriceHistories.RemoveRange(priceHistories);
